app/requirements.txt
textblob
pandas
numpy 
pyspark
pymongo 
pymongo[srv] 
tweepy 
matplotlib 
plotly 
kafka-python 
streamlit
wordcloud
findspark
flask